integrated sugiyama rendering logic into petgraph_example_page
Enough for today. Need to format the text of the nodes and also figure out why the sugiyama layout doesn't always yield the game graph. Maybe the id's a somehow broken. Or something doesn't work properly when it's being calculated in the frontend. The types of the nodes are identical (pre- and post-layout), so we could calculate the graph on the server